+++ include/ylt/reflection/member_count.hpp
@@ -4,6 +4,7 @@
 #include <tuple>
 #include <type_traits>
 #include <vector>
+#include <ylt/util/expected.hpp>
 
 namespace ylt::reflection {
 namespace internal {
@@ -25,6 +26,37 @@ struct tuple_size_impl<
 
 template <typename T>
 constexpr bool tuple_size = tuple_size_impl<T>::value;
+#endif
+
+#if __cpp_concepts >= 201907L
+  template <typename Type>
+  concept expected = requires(Type e) {
+    typename remove_cvref_t<Type>::value_type;
+    typename remove_cvref_t<Type>::error_type;
+    typename remove_cvref_t<Type>::unexpected_type;
+    e.has_value();
+    e.error();
+    requires std::is_same_v<void,
+                            typename remove_cvref_t<Type>::value_type> ||
+        requires(Type e) {
+      e.value();
+    };
+  };
+#else
+  template <typename T, typename = void>
+  struct expected_impl : std::false_type {};
+
+  template <typename T>
+  struct expected_impl<T, std::void_t<
+    typename remove_cvref_t<T>::value_type,
+    typename remove_cvref_t<T>::error_type,
+    typename remove_cvref_t<T>::unexpected_type,
+    decltype(std::declval<T>().has_value()),
+    decltype(std::declval<T>().error())>> 
+      : std::true_type {};
+    //TODO: check e.value()
+  template <typename T>
+  constexpr bool expected = expected_impl<T>::value;
 #endif
 
 #if __cpp_concepts >= 201907L
